refactor(coding): reduz complexidade cognitiva do QuestionsPanel (#391)
Extrai 4 hooks (useOutOfScopeState, useScrollToRevealedField,
useQuestionReorder, useQuestionValidation) e 2 sub-componentes
(SortableQuestion, SubmitBar) do QuestionsPanel.tsx, cada um mapeado a
uma responsabilidade isolada. Cognitive complexity da função principal
cai de 37 (HIGH) para 20 (moderate), arquivo de 412 para 200 linhas.

AssignedCodingViewProps passa a estender QuestionsPanelProps (exportada)
em vez de redeclarar os 10 campos repassados sem transformação,
resolvendo o clone-family apontado pelo fallow (dup:aa9af5e0).

Sem mudança de comportamento — QuestionsPanel.test.tsx (único teste que
exercita o componente de verdade) passa sem alteração; BrowseDocCoder.test.tsx
e CodingPage.browse.test.tsx (mockam QuestionsPanel inteiro) também.

Closes #391
