+++ frontend/src/components/coding/AssignedCodingView.tsx
@@ -6,11 +6,10 @@ import {
   ResizableHandle,
 } from "@/components/ui/resizable";
 import { DocumentReader } from "./DocumentReader";
-import { QuestionsPanel, type OutOfScopeConfig } from "./QuestionsPanel";
+import { QuestionsPanel, type QuestionsPanelProps } from "./QuestionsPanel";
 import { FullscreenNav } from "./FullscreenNav";
-import type { PydanticField } from "@/lib/types";
 
-interface AssignedCodingViewProps {
+interface AssignedCodingViewProps extends QuestionsPanelProps {
   docId: string;
   text: string;
   title: string;
@@ -19,16 +18,6 @@ interface AssignedCodingViewProps {
   isFullscreen: boolean;
   onNavigate: (index: number) => void;
   onExitFullscreen: () => void;
-  fields: PydanticField[];
-  answers: Record<string, unknown>;
-  onAnswer: (fieldName: string, value: unknown) => void;
-  onSubmit: () => void;
-  submitting: boolean;
-  notes: string;
-  onNotesChange: (notes: string) => void;
-  readOnly: boolean;
-  onReorder: (newOrder: string[]) => void;
-  outOfScope?: OutOfScopeConfig;
 }
 
 /** Painel de codificação do modo Atribuídos (leitor + perguntas). */
